+++ apps/faas-demo/package.json
@@ -16,13 +16,13 @@
     "build": "webpack --mode=production --node-env=production",
     "build:dev": "webpack --mode=development",
     "watch": "webpack --watch",
-    "generator:type": "convert-types"
+    "generator:type": "mincloud-cli type"
   },
   "bugs": {
     "url": "https://github.com/anran758/mincloud/issues"
   },
   "dependencies": {
-    "@mincloud/convert-types": "*",
+    "@mincloud/cli": "*",
     "@mincloud/io": "*"
   },
   "devDependencies": {
