+++ packages/cli/bin/cli.js
