+++ packages/cli/src/convert-types/schema.d.ts
